requirements.txt
streamlit>=1.38.0
python-pptx>=0.6.23